+++ kutana/backends/vkontakte/__init__.py
@@ -1,0 +1,4 @@
+from .extensions import VkontaktePluginExtension
+from .backend import Vkontakte
+
+__all__ = ["VkontaktePluginExtension", "Vkontakte"]